+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 In this file I'll try to mention major changes to the configuration. 
 
+2025-11-18
+- flake: added latest ulauncher for t14 (for now)
+
 2025-11-17
 - module: added hyprland wayland compositor to replace Gnome DE.